% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgmfiles-package.r
\docType{package}
\name{bgmfiles}
\alias{bgmfiles}
\alias{bgmfiles-package}
\title{bgmfiles.}
\description{
Raw data files are maintained here: https://github.com/AustralianAntarcticDivision/box-geometry-models.git
}
\details{
This package contains a simple workflow to update from that source repo and install the raw files into extdata/ of the installed package.
}

